chore: rename consensys -> linea
